+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [11.1.0] (IN PROGRESS)
 
 * Replace `moment` usage with `dayjs`. (UIEH-1407)
+* Show a loading indicator under Package Titles list when package titles are updating. (UIEH-1471)
 
 ## [11.0.1] (https://github.com/folio-org/ui-eholdings/tree/v11.0.1) (2025-04-16)