+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Package-Title MCL: indicate HLM title visibility (ie Hidden) in status column. (UIEH-1464)
 * Package Details Record: Title List Search within: Make reset action work. (UIEH-1466)
 * Make "erm" an optional interface in eHoldings. (UIEH-1476)
+* include global permissions in package.json base permissions. (UIEH-1481)
 
 ## [11.0.1] (https://github.com/folio-org/ui-eholdings/tree/v11.0.1) (2025-04-16)